Refactored issuance and dashboard modules, transfered Transaction and Document details to feature-dashboard module.
