+++ dashboard-feature/src/main/java/eu/europa/ec/dashboardfeature/interactor/DocumentDetailsInteractor.kt
@@ -14,7 +14,7 @@
  * governing permissions and limitations under the Licence.
  */
 
-package eu.europa.ec.issuancefeature.interactor.document
+package eu.europa.ec.dashboardfeature.interactor
 
 import eu.europa.ec.businesslogic.extension.safeAsync
 import eu.europa.ec.commonfeature.ui.document_details.domain.DocumentDetailsDomain
