+++ dashboard-feature/src/main/java/eu/europa/ec/dashboardfeature/router/Graph.kt
@@ -18,15 +18,21 @@ package eu.europa.ec.dashboardfeature.router
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
+import androidx.navigation.NavType
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.navigation
+import androidx.navigation.navArgument
 import androidx.navigation.navDeepLink
 import eu.europa.ec.dashboardfeature.BuildConfig
 import eu.europa.ec.dashboardfeature.ui.dashboard.DashboardScreen
+import eu.europa.ec.dashboardfeature.ui.documents.detail.DocumentDetailsScreen
 import eu.europa.ec.dashboardfeature.ui.sign.DocumentSignScreen
+import eu.europa.ec.dashboardfeature.ui.transactions.detail.TransactionDetailsScreen
 import eu.europa.ec.uilogic.navigation.DashboardScreens
 import eu.europa.ec.uilogic.navigation.ModuleRoute
+import org.koin.androidx.compose.getViewModel
 import org.koin.androidx.compose.koinViewModel
+import org.koin.core.parameter.parametersOf
 
 fun NavGraphBuilder.featureDashboardGraph(navController: NavController) {
     navigation(
@@ -55,6 +61,58 @@ fun NavGraphBuilder.featureDashboardGraph(navController: NavController) {
             route = DashboardScreens.SignDocument.screenRoute
         ) {
             DocumentSignScreen(navController, koinViewModel())
+        }
+
+        composable(
+            route = DashboardScreens.DocumentDetails.screenRoute,
+            deepLinks = listOf(
+                navDeepLink {
+                    uriPattern =
+                        BuildConfig.DEEPLINK + DashboardScreens.DocumentDetails.screenRoute
+                }
+            ),
+            arguments = listOf(
+                navArgument("documentId") {
+                    type = NavType.StringType
+                },
+            )
+        ) {
+            DocumentDetailsScreen(
+                navController,
+                getViewModel(
+                    parameters = {
+                        parametersOf(
+                            it.arguments?.getString("documentId").orEmpty(),
+                        )
+                    }
+                )
+            )
+        }
+
+        composable(
+            route = DashboardScreens.TransactionDetails.screenRoute,
+            deepLinks = listOf(
+                navDeepLink {
+                    uriPattern =
+                        BuildConfig.DEEPLINK + DashboardScreens.TransactionDetails.screenRoute
+                }
+            ),
+            arguments = listOf(
+                navArgument("transactionId") {
+                    type = NavType.StringType
+                },
+            )
+        ) {
+            TransactionDetailsScreen(
+                navController,
+                getViewModel(
+                    parameters = {
+                        parametersOf(
+                            it.arguments?.getString("transactionId").orEmpty(),
+                        )
+                    }
+                )
+            )
         }
     }
 }
