+++ dashboard-feature/src/main/java/eu/europa/ec/dashboardfeature/ui/transactions/list/TransactionsViewModel.kt
@@ -14,7 +14,7 @@
  * governing permissions and limitations under the Licence.
  */
 
-package eu.europa.ec.dashboardfeature.ui.transactions
+package eu.europa.ec.dashboardfeature.ui.transactions.list
 
 import androidx.lifecycle.viewModelScope
 import eu.europa.ec.businesslogic.util.toLocalDate
@@ -39,7 +39,6 @@ import eu.europa.ec.uilogic.mvi.ViewEvent
 import eu.europa.ec.uilogic.mvi.ViewSideEffect
 import eu.europa.ec.uilogic.mvi.ViewState
 import eu.europa.ec.uilogic.navigation.DashboardScreens
-import eu.europa.ec.uilogic.navigation.IssuanceScreens
 import eu.europa.ec.uilogic.navigation.helper.generateComposableArguments
 import eu.europa.ec.uilogic.navigation.helper.generateComposableNavigationLink
 import kotlinx.coroutines.launch
@@ -363,7 +362,7 @@ class TransactionsViewModel(
         setEffect {
             Effect.Navigation.SwitchScreen(
                 screenRoute = generateComposableNavigationLink(
-                    screen = IssuanceScreens.TransactionDetails,
+                    screen = DashboardScreens.TransactionDetails,
                     arguments = generateComposableArguments(
                         mapOf(
                             "transactionId" to transactionId
