+++ gradle/libs.versions.toml
@@ -1,7 +1,7 @@
 [versions]
 accompanist = "0.34.0"
 androidDesugarJdkLibs = "2.1.5"
-androidGradlePlugin = "8.9.1"
+androidGradlePlugin = "8.9.2"
 androidxActivity = "1.10.1"
 androidxAppCompat = "1.7.0"
 androidxBrowser = "1.8.0"
@@ -36,7 +36,7 @@ kotlinxCoroutines = "1.9.0"
 kotlinxDatetime = "0.6.2"
 kotlinxSerializationJson = "1.8.0"
 ksp = "2.0.21-1.0.27"
-lint = "31.9.1"
+lint = "31.9.2"
 okhttp = "4.12.0"
 protobuf = "3.24.0"
 protobufPlugin = "0.9.4"
