+++ issuance-feature/src/main/java/eu/europa/ec/issuancefeature/interactor/DocumentIssuanceSuccessInteractor.kt
@@ -14,7 +14,7 @@
  * governing permissions and limitations under the Licence.
  */
 
-package eu.europa.ec.issuancefeature.interactor.document
+package eu.europa.ec.issuancefeature.interactor
 
 import eu.europa.ec.businesslogic.extension.safeAsync
 import eu.europa.ec.commonfeature.extensions.toExpandableListItems
