+++ issuance-feature/src/main/java/eu/europa/ec/issuancefeature/router/Graph.kt
@@ -30,10 +30,8 @@ import eu.europa.ec.commonfeature.config.OfferUiConfig
 import eu.europa.ec.issuancefeature.BuildConfig
 import eu.europa.ec.issuancefeature.ui.document.add.AddDocumentScreen
 import eu.europa.ec.issuancefeature.ui.document.code.DocumentOfferCodeScreen
-import eu.europa.ec.issuancefeature.ui.document.details.DocumentDetailsScreen
 import eu.europa.ec.issuancefeature.ui.document.offer.DocumentOfferScreen
 import eu.europa.ec.issuancefeature.ui.document.success.DocumentIssuanceSuccessScreen
-import eu.europa.ec.issuancefeature.ui.transaction.details.TransactionDetailsScreen
 import eu.europa.ec.uilogic.navigation.IssuanceScreens
 import eu.europa.ec.uilogic.navigation.ModuleRoute
 import org.koin.androidx.compose.getViewModel
@@ -67,60 +65,6 @@ fun NavGraphBuilder.featureIssuanceGraph(navController: NavController) {
                             IssuanceFlowUiConfig.fromString(
                                 it.arguments?.getString("flowType").orEmpty()
                             ),
-                        )
-                    }
-                )
-            )
-        }
-
-        // Document Details
-        composable(
-            route = IssuanceScreens.DocumentDetails.screenRoute,
-            deepLinks = listOf(
-                navDeepLink {
-                    uriPattern =
-                        BuildConfig.DEEPLINK + IssuanceScreens.DocumentDetails.screenRoute
-                }
-            ),
-            arguments = listOf(
-                navArgument("documentId") {
-                    type = NavType.StringType
-                },
-            )
-        ) {
-            DocumentDetailsScreen(
-                navController,
-                getViewModel(
-                    parameters = {
-                        parametersOf(
-                            it.arguments?.getString("documentId").orEmpty(),
-                        )
-                    }
-                )
-            )
-        }
-
-        // Transaction Details
-        composable(
-            route = IssuanceScreens.TransactionDetails.screenRoute,
-            deepLinks = listOf(
-                navDeepLink {
-                    uriPattern =
-                        BuildConfig.DEEPLINK + IssuanceScreens.TransactionDetails.screenRoute
-                }
-            ),
-            arguments = listOf(
-                navArgument("transactionId") {
-                    type = NavType.StringType
-                },
-            )
-        ) {
-            TransactionDetailsScreen(
-                navController,
-                getViewModel(
-                    parameters = {
-                        parametersOf(
-                            it.arguments?.getString("transactionId").orEmpty(),
                         )
                     }
                 )
