+++ issuance-feature/src/main/java/eu/europa/ec/issuancefeature/ui/document/code/DocumentOfferCodeViewModel.kt
@@ -21,8 +21,8 @@ import androidx.lifecycle.viewModelScope
 import eu.europa.ec.commonfeature.config.IssuanceSuccessUiConfig
 import eu.europa.ec.commonfeature.config.OfferCodeUiConfig
 import eu.europa.ec.eudi.wallet.document.DocumentId
-import eu.europa.ec.issuancefeature.interactor.document.DocumentOfferInteractor
-import eu.europa.ec.issuancefeature.interactor.document.IssueDocumentsInteractorPartialState
+import eu.europa.ec.issuancefeature.interactor.DocumentOfferInteractor
+import eu.europa.ec.issuancefeature.interactor.IssueDocumentsInteractorPartialState
 import eu.europa.ec.resourceslogic.R
 import eu.europa.ec.resourceslogic.provider.ResourceProvider
 import eu.europa.ec.uilogic.component.content.ContentErrorConfig
